langchain/custom-llms-and-embeddings/pyproject.toml
[tool.poetry]
name = "custom-llms"
version = "0.1.0"
description = ""
authors = ["Helton Carlos de Souza <heltoncarlossouza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ctransformers = "^0.2.27"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
gpt4all = "^2.0.2"
pypdf = "^3.17.1"
langchain = "^0.0.336"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
